@@ -33,7 +33,7 @@
     "ts-jest": "^19.0.10",
     "tslint": "^5.1.0",
     "tslint-react": "^2.6.0",
-    "typescript": "^2.3.2"
+    "typescript": "^2.7.2"
   },
   "pre-commit": [],
   "jest": {
