@@ -1,12 +1,14 @@
-import isFunction = require('lodash/isFunction');
+import isFunction from 'lodash/isFunction';
 export enum DECORATE_TARGET {
   UN_KNOW = -1,
   CLASS,
   METHOD
 }
 
-// tslint:disable-next-line:ban-types
-export function decorateTarget(target: Function, descriptor?: PropertyDescriptor) {
+export function decorateTarget(
+  target: Function,
+  descriptor?: PropertyDescriptor
+) {
   if (descriptor && isFunction(descriptor.value)) {
     return DECORATE_TARGET.METHOD;
   } else if (isFunction(target)) {
