@@ -1,5 +1,6 @@
-import decorateTarget from '../../src/utils/decorateTarget';
-import { DECORATE_TARGET } from '../../src/utils/decorateTarget';
+import decorateTarget, {
+  DECORATE_TARGET
+} from '../../src/utils/decorateTarget';
 
 let classArgs: any = null;
 let staticMethodArgs: any = null;
@@ -8,8 +9,7 @@ let protertyArgs: any = null;
 
 @setClassArgs
 class A {
-  @setProtertyArgs
-  public name: string;
+  @setProtertyArgs public name!: string;
   @setStaticMethodArgs
   public static Say() {
     return 'static say';
@@ -39,8 +39,14 @@ function setProtertyArgs(...args: any[]) {
 describe('test utils/decorateTarget', () => {
   test('descorate target', () => {
     expect(decorateTarget(classArgs[0])).toBe(DECORATE_TARGET.CLASS);
-    expect(decorateTarget(staticMethodArgs[0], staticMethodArgs[2])).toBe(DECORATE_TARGET.METHOD);
-    expect(decorateTarget(methodArgs[0], methodArgs[2])).toBe(DECORATE_TARGET.METHOD);
-    expect(decorateTarget(protertyArgs[0], protertyArgs[2])).toBe(DECORATE_TARGET.UN_KNOW);
+    expect(decorateTarget(staticMethodArgs[0], staticMethodArgs[2])).toBe(
+      DECORATE_TARGET.METHOD
+    );
+    expect(decorateTarget(methodArgs[0], methodArgs[2])).toBe(
+      DECORATE_TARGET.METHOD
+    );
+    expect(decorateTarget(protertyArgs[0], protertyArgs[2])).toBe(
+      DECORATE_TARGET.UN_KNOW
+    );
   });
 });
